\name{mapa_peru}
\alias{mapa_peru}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Mapa of Peru}
\description{
generate a base map from Peru
}
\usage{
mapa_peru()
}
%- maybe also 'usage' for other objects documented here.

\details{
using \code{kali package} and generate a base map from Peru. you need install kali from https://github.com/roliveros-ramos
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
mapa_peru()
points(dataLand$lon, dataLand$lat)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
